+++ BlueToot/View Hierarchy/Timeline/DataAdapters/StatusDataProvider.swift
@@ -1,0 +1,27 @@
+//
+//  StatusDataProvider.swift
+//  BlueToot
+//
+//  Created by Kevin Clark on 9/3/18.
+//  Copyright © 2018 Kevin Clark. All rights reserved.
+//
+
+import UIKit
+import MastodonKit
+
+class StatusDataProvider: ArrayDataProvider {
+  var items: [Any] = []
+
+  func loadData(completion: (() -> Void)?) {
+    let homeRequest = Timelines.home(range: .limit(100))
+
+    Mastodon.client.run(homeRequest) { result in
+      guard let statuses = result.value else { return }
+
+      DispatchQueue.main.async {
+        self.items = statuses
+        completion?()
+      }
+    }
+  }
+}
